Journals and References/YOLO/scripts/tests/requirements.txt
# Test Requirements for Auth Service Integration Tests
# Smart Office/Home Surveillance System

# Testing framework
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0  # For parallel test execution

# HTTP client
httpx==0.26.0

# Database
psycopg2-binary==2.9.9

# JWT handling
PyJWT==2.8.0
cryptography==42.0.0

# Async support
asyncio==3.4.3

# Test utilities
faker==22.6.0  # For generating test data
freezegun==1.4.0  # For mocking time

# Code coverage
coverage==7.4.0
